Created basic structure for the new website design
