+++ next.config.js
@@ -1,10 +1,11 @@
 const withPlugins = require('next-compose-plugins');
 const withFonts = require('next-fonts');
 const withMDX = require('@next/mdx')();
+const withSass = require('@zeit/next-sass');
 const OptimizeCSSAssetsPlugin = require('optimize-css-assets-webpack-plugin');
 const withCSS = require('./src/utils/withCss'); // FIXME Own implementation used until this is fixed: https://github.com/zeit/next-plugins/issues/526
 
-module.exports = withPlugins([ withCSS, withFonts, withMDX ], {
+module.exports = withPlugins([ withCSS, withSass, withFonts, withMDX ], {
   enableSvg: true,
   exportTrailingSlash: true, // Makes pages to be exported as index.html files
   env: { // Any of these vars can be accessed as process.env.{name}
