+++ src/assets/sass/main.scss
@@ -1,0 +1,5 @@
+@import "theme/theme";
+
+.logo-icon {
+  width: 40px;
+}
