+++ src/components/Footer.tsx
@@ -2,29 +2,13 @@ import React, { FunctionComponent } from 'react';
 import { ExternalLink } from 'react-external-link';
 
 const Footer: FunctionComponent = () => (
-  <footer id="footer">
-    <ul className="icons">
-      <li>
-        <ExternalLink href="https://github.com/shlinkio/shlink" className="icon fa-github">
-          <span className="label">Github</span>
-        </ExternalLink>
-      </li>
-      <li>
-        <ExternalLink href="https://twitter.com/shlinkio" className="icon fa-twitter">
-          <span className="label">Twitter</span>
-        </ExternalLink>
-      </li>
-      <li>
-        <ExternalLink href={process.env.donateUrl || ''} className="icon fa-paypal">
-          <span className="label">Donate</span>
-        </ExternalLink>
-      </li>
-    </ul>
-    <p className="copyright">
-      &copy; {new Date().getFullYear()} Shlink. With <span className="fa fa-heart" title="love" /> by{' '}
-      <ExternalLink href="https://www.alejandrocelaya.com">Alejandro Celaya</ExternalLink> -
-      Template from <a href="http://html5up.net">HTML5 UP</a>
-    </p>
+  <footer className="footer">
+    <div className="container text-center py-5">
+      <p className="copyright">
+        &copy; {new Date().getFullYear()} Shlink. With <span className="fa fa-heart" title="love" /> by{' '}
+        <ExternalLink href="https://www.alejandrocelaya.com">Alejandro Celaya</ExternalLink>
+      </p>
+    </div>
   </footer>
 );
