+++ src/components/Menu.tsx
@@ -1,67 +1,84 @@
-import React, { useState, useEffect, FunctionComponent } from 'react';
+import React, { FunctionComponent, useState } from 'react';
 import Link from 'next/link';
-import { useRouter } from 'next/router';
 import classNames from 'classnames';
 import { ExternalLink } from 'react-external-link';
+import { Collapse, NavbarToggler } from 'reactstrap';
 import './Menu.css';
 
-// FIXME Make window object to be passed as prop
 const Menu: FunctionComponent = () => {
-  const [ active, setActive ] = useState(false);
-  const [ show, setShowSubmenu ] = useState(false);
-  const { pathname: currentPage } = useRouter();
-
-  useEffect(() => {
-    window.onscroll = () => setActive(window.scrollY >= 20);
-
-    return () => {
-      window.onscroll = null;
-    };
-  }, []);
+  const [ collapsed, setCollapsed ] = useState(true);
+  const toggleCollapsed = () => setCollapsed(!collapsed);
 
   return (
-    <nav id="navbar" className={classNames({ active })}>
-      <Link href="/">
-        <a className="logo">
-          <img src="/images/shlink-logo-white.png" />&nbsp;&nbsp;Shlink
-        </a>
-      </Link>
+    <header className="header fixed-top">
+      <div className="container-fluid position-relative">
+        <nav className="navbar navbar-expand-lg">
+          <div className="site-logo">
+            <Link href="/">
+              <a className="navbar-brand">
+                <img
+                  className="logo-icon mr-2"
+                  src="/images/shlink-logo-blue.svg"
+                  alt="logo"
+                />
+                <span className="logo-text">Shlink</span>
+              </a>
+            </Link>
+          </div>
 
-      <button className="menu-collapser" onClick={() => setShowSubmenu(!show)}><i className="fa fa-bars" /></button>
+          <NavbarToggler className={classNames({ collapsed })} onClick={toggleCollapsed}>
+            <span />
+            <span />
+            <span />
+          </NavbarToggler>
 
-      <ul className={classNames('menu main-menu', { show })}>
-        <li className={currentPage.startsWith('/features') ? 'active' : ''}>
-          <Link href="/features">
-            <a><i className="fa fa-bolt" />&nbsp;&nbsp;Features</a>
-          </Link>
-        </li>
-        <li className={currentPage.startsWith('/documentation') ? 'active' : ''}>
-          <Link href="/documentation">
-            <a><i className="fa fa-book" />&nbsp;&nbsp;Documentation</a>
-          </Link>
-        </li>
-        <li className={currentPage.startsWith('/command-line-interface') ? 'active' : ''}>
-          <Link href="/command-line-interface">
-            <a><i className="fa fa-terminal" />&nbsp;&nbsp;CLI</a>
-          </Link>
-        </li>
-        <li className={currentPage.startsWith('/api-docs') ? 'active' : ''}>
-          <Link href="/api-docs">
-            <a><i className="fa fa-server" />&nbsp;&nbsp;API Docs</a>
-          </Link>
-        </li>
-        <li className={currentPage.startsWith('/apps') ? 'active' : ''}>
-          <Link href="/apps">
-            <a><i className="fa fa-rocket" />&nbsp;&nbsp;Apps</a>
-          </Link>
-        </li>
-        <li>
-          <ExternalLink href={process.env.donateUrl || ''}>
-            <i className="fa fa-paypal" />&nbsp;&nbsp;Donate
-          </ExternalLink>
-        </li>
-      </ul>
-    </nav>
+          <Collapse isOpen={!collapsed} navbar>
+            <ul className="social-list list-inline mt-3 mt-lg-0 mb-lg-0 d-flex ml-lg-5 mr-lg-5">
+              <li className="list-inline-item">
+                <ExternalLink href="https://github.com/shlinkio/shlink">
+                  <i className="fa fa-github fa-fw" />
+                </ExternalLink>
+              </li>
+              <li className="list-inline-item">
+                <ExternalLink href="https://twitter.com/shlinkio"><i className="fa fa-twitter fa-fw" /></ExternalLink>
+              </li>
+              <li className="list-inline-item">
+                <ExternalLink href={process.env.donateUrl || ''}><i className="fa fa-paypal fa-fw" /></ExternalLink>
+              </li>
+            </ul>
+
+            <ul className="navbar-nav ml-lg-auto">
+              <li className="nav-item mr-lg-4">
+                <Link href="/features">
+                  <a className="nav-link">Features</a>
+                </Link>
+              </li>
+              <li className="nav-item mr-lg-4">
+                <Link href="/documentation">
+                  <a className="nav-link">Docs</a>
+                </Link>
+              </li>
+              <li className="nav-item mr-lg-4">
+                <Link href="/command-line-interface">
+                  <a className="nav-link">CLI</a>
+                </Link>
+              </li>
+              <li className="nav-item mr-lg-4">
+                <Link href="/api-docs">
+                  <a className="nav-link">API Docs</a>
+                </Link>
+              </li>
+              <li className="nav-item mr-lg-0">
+                <Link href="/apps">
+                  <a className="nav-link">Apps</a>
+                </Link>
+              </li>
+            </ul>
+          </Collapse>
+        </nav>
+
+      </div>
+    </header>
   );
 };
