+++ src/pages/_app.ts
@@ -4,7 +4,6 @@ import withGA from 'next-ga';
 
 import 'highlight.js/styles/atom-one-light.css';
 import 'font-awesome/css/font-awesome.css';
-import '../assets/css/vendor.css';
-import '../assets/css/main.css';
+import '../assets/sass/main.scss';
 
 export default withGA(process.env.ga, Router)(App);
